pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "Samuel Miranda",email = "samuelmirandasamt@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "realmate_challenge", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
